+++ src/api/api.ts
@@ -13,7 +13,7 @@ import Toot, { SerializableToot, earliestTootedAt, mostRecentTootedAt, sortByCre
 import UserData from "./user_data";
 import { ageString, mostRecent, quotedISOFmt, subtractSeconds, timelineCutoffAt } from "../helpers/time_helpers";
 import { ApiMutex, MastodonApiObject, MastodonObjWithID, MastodonTag, StatusList, CacheKey } from "../types";
-import { Config, MIN_RECORDS_FOR_FEATURE_SCORING } from "../config";
+import { config, MIN_RECORDS_FOR_FEATURE_SCORING } from "../config";
 import { findMinMaxId, truncateToConfiguredLength } from "../helpers/collection_helpers";
 import { lockExecution, logAndThrowError, traceLog } from '../helpers/log_helpers';
 import { repairTag } from "./objects/tag";
@@ -76,7 +76,7 @@ export default class MastoApi {
     user: Account;
     userData?: UserData;  // Save UserData in the API object to avoid polling local storage over and over
     private mutexes: ApiMutex;  // Mutexes for blocking singleton requests (e.g. followed accounts)
-    private requestSemphore = new Semaphore(Config.api.maxConcurrentRequestsInitial); // Limit concurrency of search & tag requests
+    private requestSemphore = new Semaphore(config.api.maxConcurrentRequestsInitial); // Limit concurrency of search & tag requests
 
     // URL for tag on the user's homeserver
     tagUrl = (tag: MastodonTag | string) => `${this.endpointURL(TAGS)}/${typeof tag === "string" ? tag : tag.name}`;
@@ -129,7 +129,7 @@ export default class MastoApi {
         } else {
             // Look back additional lookbackForUpdatesMinutes minutes to catch new updates and edits to toots
             const maxTootedAt = mostRecentTootedAt(homeTimelineToots);
-            const lookbackSeconds = Config.api.data[CacheKey.HOME_TIMELINE]?.lookbackForUpdatesMinutes! * 60;
+            const lookbackSeconds = config.api.data[CacheKey.HOME_TIMELINE]?.lookbackForUpdatesMinutes! * 60;
             cutoffAt = maxTootedAt ? subtractSeconds(maxTootedAt, lookbackSeconds) : timelineCutoffAt();
             cutoffAt = mostRecent(timelineCutoffAt(), cutoffAt)!;
             console.debug(`${logPrefix} maxTootedAt: ${quotedISOFmt(maxTootedAt)}, maxId: ${maxId}, cutoffAt: ${quotedISOFmt(cutoffAt)}`);
@@ -170,7 +170,7 @@ export default class MastoApi {
         let msg = `${logPrefix} Fetched ${allNewToots.length} new toots ${ageString(startedAt)} (${oldestTootStr}`;
         console.debug(`${msg}, home feed has ${homeTimelineToots.length} toots)`);
         homeTimelineToots = sortByCreatedAt(homeTimelineToots); // TODO: should we sort by score?
-        homeTimelineToots = truncateToConfiguredLength(homeTimelineToots, Config.toots.maxCachedTimelineToots, logPrefix);
+        homeTimelineToots = truncateToConfiguredLength(homeTimelineToots, config.toots.maxCachedTimelineToots, logPrefix);
         await Storage.set(cacheKey, homeTimelineToots);
         return homeTimelineToots;
     }
@@ -307,7 +307,7 @@ export default class MastoApi {
     // Get latest toots for a given tag using both the Search API and tag timeline API.
     // The two APIs give results with surprising little overlap (~80% of toots are unique)
     async getStatusesForTag(tag: MastodonTag, numToots?: number): Promise<mastodon.v1.Status[]> {
-        numToots ||= Config.trending.tags.numTootsPerTag;
+        numToots ||= config.trending.tags.numTootsPerTag;
         const startedAt = new Date();
 
         const tagToots = await Promise.all([
@@ -341,7 +341,7 @@ export default class MastoApi {
     // See https://docs.joinmastodon.org/methods/timelines/#tag
     // TODO: we could use the min_id param to avoid redundancy and extra work reprocessing the same toots
     async hashtagTimelineToots(tag: MastodonTag, maxRecords?: number): Promise<Toot[]> {
-        maxRecords = maxRecords || Config.api.defaultRecordsPerPage;
+        maxRecords = maxRecords || config.api.defaultRecordsPerPage;
         const logPrefix = `[hashtagTimelineToots("#${tag.name}")]`;
         const releaseSemaphore = await lockExecution(this.requestSemphore, logPrefix);
         const startedAt = new Date();
@@ -391,7 +391,7 @@ export default class MastoApi {
     //   - searchString:  the string to search for
     //   - maxRecords:    the maximum number of records to fetch
     async searchForToots(searchStr: string, maxRecords?: number): Promise<mastodon.v1.Status[]> {
-        maxRecords = maxRecords || Config.api.defaultRecordsPerPage;
+        maxRecords = maxRecords || config.api.defaultRecordsPerPage;
         let logPrefix = `[API searchForToots("${searchStr}")]`;
         const releaseSemaphore = await lockExecution(this.requestSemphore, logPrefix);
         const query: mastodon.rest.v1.SearchParams = {limit: maxRecords, q: searchStr, type: STATUSES};
@@ -432,9 +432,9 @@ export default class MastoApi {
         if (moar && (skipCache || maxId)) console.warn(`${logPfx} skipCache=true AND moar or maxId set`);
 
         // Parse params and set defaults
-        const requestDefaults = Config.api.data[cacheKey];
+        const requestDefaults = config.api.data[cacheKey];
         maxRecords ??= requestDefaults?.initialMaxRecords ?? MIN_RECORDS_FOR_FEATURE_SCORING;
-        const limit = Math.min(maxRecords, requestDefaults?.limit || Config.api.defaultRecordsPerPage);  // max records per page
+        const limit = Math.min(maxRecords, requestDefaults?.limit || config.api.defaultRecordsPerPage);  // max records per page
         breakIf ??= DEFAULT_BREAK_IF;
         let minId: string | undefined; // Used for incremental loading when data is stale (if supported)