+++ src/api/moar_data_poller.ts
@@ -6,7 +6,7 @@ import { Mutex } from 'async-mutex';
 
 import MastoApi from "../api/api";
 import { ageString } from '../helpers/time_helpers';
-import { Config } from "../config";
+import { config } from "../config";
 import { lockExecution } from '../helpers/log_helpers';
 
 export const GET_MOAR_DATA = "getMoarData()";
@@ -19,7 +19,7 @@ const MOAR_MUTEX = new Mutex();
 // TODO: Add followed accounts?  for people who follow a lot?
 export async function getMoarData(): Promise<boolean> {
     console.log(`${MOAR_DATA_PREFIX} triggered by timer...`);
-    const maxRecordsForFeatureScoring = Config.api.maxRecordsForFeatureScoring;
+    const maxRecordsForFeatureScoring = config.api.maxRecordsForFeatureScoring;
     const startedAt = new Date();
     const releaseMutex = await lockExecution(MOAR_MUTEX, GET_MOAR_DATA);