+++ src/scorer/feed/diversity_feed_scorer.ts
@@ -4,7 +4,7 @@
  */
 import FeedScorer from "../feed_scorer";
 import Toot, { sortByCreatedAt } from '../../api/objects/toot';
-import { Config } from "../../config";
+import { config } from "../../config";
 import { decrementCount, incrementCount } from "../../helpers/collection_helpers";
 import { ScoreName, StringNumberDict } from "../../types";
 import { traceLog } from "../../helpers/log_helpers";
@@ -74,7 +74,7 @@ export default class DiversityFeedScorer extends FeedScorer {
 
                     if (toot.account.isFollowed || toot.reblog?.account.isFollowed) {
                         // if (toot.trendingTags?.length) traceLog(`${this.logPrefix()} Not penalizing followed toot:`, toot.realToot().describe());
-                    } else if (tootsWithTagScoredSoFar[tag.name] > Config.scoring.minTrendingTagTootsForPenalty) {
+                    } else if (tootsWithTagScoredSoFar[tag.name] > config.scoring.minTrendingTagTootsForPenalty) {
                         // TODO: this suddenly applies a massive penalty to the first toot beyond the threshold
                         // ...but only apply the penalty after MIN_TRENDING_TAGS_FOR_PENALTY toots have been passed over
                         scores[toot.uri] -= trendingTagPenalty[tag.name] || 0;